Partially implement NetCdfWeatherVariableRasterSlice in ArrayWeatherVariableRasterSlice
