+++ src/main/kotlin/com/luca009/imker/server/parser/NetCdfWeatherVariableRasterSlice.kt
@@ -13,69 +13,39 @@ class NetCdfWeatherVariableRasterSlice(
     private val data: Array<*>,
     ucarDimensions: List<Dimension>,
     isLatLon: Boolean
-) : WeatherVariableRasterSlice(unit, dataType) {
+) : ArrayWeatherVariableRasterSlice(
+    unit,
+    dataType,
+    data,
+    ucarDimensions.mapNotNull {
+        val enum = if (isLatLon) {
+            // If the coordinate system is latlon, then map the latitude and longitude variables to X and Y respectively
+            when (it.name.lowercase()) {
+                "latitude" -> WeatherVariableRasterDimensionType.X
+                "longitude" -> WeatherVariableRasterDimensionType.Y
+                else -> getDimensionTypeFromName(it.name)
+            }
+        } else {
+            getDimensionTypeFromName(it.name)
+        }
+
+        requireNotNull(enum) {
+            return@mapNotNull null
+        }
+
+        Pair(
+            enum,
+            WeatherVariableRasterDimension(
+                it.length
+            )
+        )
+    }.toMap()
+) {
+
+    private companion object {
+        fun getDimensionTypeFromName(name: String): WeatherVariableRasterDimensionType? = WeatherVariableRasterDimensionType.values().firstOrNull { it.name.equals(name, true) }
+    }
+
     override val size: Int
         get() = data.size
-
-    override val dimensions: Map<WeatherVariableRasterDimensionType, WeatherVariableRasterDimension>
-    private val splitFrequencies: List<Int>
-
-    init {
-        dimensions = ucarDimensions.mapNotNull {
-            val enum = if (isLatLon) {
-                // If the coordinate system is latlon, then map the latitude and longitude variables to X and Y respectively
-                when (it.name.lowercase()) {
-                    "latitude" -> WeatherVariableRasterDimensionType.X
-                    "longitude" -> WeatherVariableRasterDimensionType.Y
-                    else -> getDimensionTypeFromName(it.name)
-                }
-            } else {
-                getDimensionTypeFromName(it.name)
-            }
-
-            requireNotNull(enum) {
-                return@mapNotNull null
-            }
-
-            Pair(
-                enum,
-                WeatherVariableRasterDimension(
-                    it.length
-                )
-            )
-        }.toMap()
-
-        splitFrequencies = dimensions.values.reversed().toList().dropLast(1).map { it.size }
-    }
-
-    private fun getDimensionTypeFromName(name: String): WeatherVariableRasterDimensionType? = WeatherVariableRasterDimensionType.values().firstOrNull { it.name.equals(name, true) }
-
-    private fun getIndex(indices: IntArray): Int {
-        return indices.mapIndexed { index, value ->
-            if (index == 0) {
-                value
-            } else {
-                value * splitFrequencies[index - 1]
-            }
-        }.sum()
-    }
-
-    override fun get(vararg indices: Int): Any? {
-        if (splitFrequencies.size + 1 != indices.size) {
-            throw IllegalArgumentException("Number of indices was out of range (expected ${splitFrequencies.size + 1}, received ${indices.size})")
-        }
-
-        val index = getIndex(indices)
-        return data[index]
-    }
-
-    override fun getOrNull(vararg indices: Int): Any? {
-        if (splitFrequencies.size + 1 != indices.size) {
-            return null
-        }
-
-        val index = getIndex(indices)
-        return data.getOrNull(index)
-    }
-
 }
